EPPlus/FormulaParsing/Excel/Functions/Math/Round.cs
/* Copyright (C) 2011  Jan Källman
 *
 * This library is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 2.1 of the License, or (at your option) any later version.

 * This library is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  
 * See the GNU Lesser General Public License for more details.
 *
 * The GNU Lesser General Public License can be viewed at http://www.opensource.org/licenses/lgpl-license.php
 * If you unfamiliar with this license or have questions about it, here is an http://www.gnu.org/licenses/gpl-faq.html
 *
 * All code and executables are provided "as is" with no warranty either express or implied. 
 * The author accepts no liability for any damage or loss of business that this product may cause.
 *
 * Code change notes:
 * 
 * Author							Change						Date
 *******************************************************************************
 * Mats Alm   		                Added		                2013-12-03
 *******************************************************************************/
using System.Collections.Generic;
using System.Linq;
using OfficeOpenXml.FormulaParsing.ExpressionGraph;
using OfficeOpenXml.Utils;

namespace OfficeOpenXml.FormulaParsing.Excel.Functions.Math
{
	public class Round : ExcelFunction
	{
		public override CompileResult Execute(IEnumerable<FunctionArgument> arguments, ParsingContext context)
		{
			if (this.ArgumentsAreValid(arguments, 2, out eErrorType argumentError) == false)
				return new CompileResult(argumentError);

			if (arguments.ElementAt(0).Value == null)
				return CreateResult(0d, DataType.Decimal);

			if (!ConvertUtil.TryParseDateObjectToOADate(arguments.ElementAt(0).Value, out double result))
				return new CompileResult(eErrorType.Value);

			if (!ConvertUtil.TryParseDateObjectToOADate(arguments.ElementAt(1).Value, out double result2))
				return new CompileResult(eErrorType.Value);

			var number = result;
			var nDigits = (int)result2;
			if (nDigits > 15)
				nDigits = 15;
			if (nDigits < 0)
			{
				nDigits *= -1;
				return CreateResult(number - (number % (System.Math.Pow(10, nDigits))), DataType.Integer);
			}
			return CreateResult(System.Math.Round(number, nDigits), DataType.Decimal);
		}
	}
}
